Combine GingaView and GingaControl classes into one
- this merges the GingaView and GingaControl classes into a single
  GingaShell class.  They no longer need to be separate and this
  greatly simplifies documenting and understanding the class
- deprecated some files marked for deprecation
