--- ginga/misc/Plot.py
@@ -1,2 +1,0 @@
-# TO BE DEPRECATED
-from ginga.gw.Plot import *
