--- ginga/misc/Widgets.py
@@ -1,2 +1,0 @@
-# TO BE DEPRECATED
-from ginga.gw.Widgets import *
